match handleCommands thanks GoldieLeGenie, ED_1T!

+++ include/System/CProc.h
@@ -35,8 +35,8 @@ public:
     void func_02014320(int);
     void func_02014408(int, int);
     void func_02014454(u32);
-    void func_02014520(int, u16, u8, u32);
-    void func_02014588(int, int, u16, u8, u32);
+    void func_02014520(int, u32, u32, s32);
+    void func_02014588(int, int, u32, u32, s32);
     void func_020145f4(int, u16, u32);
     void func_02014620(void);
     int func_02014880(int);
